wip: payment

+++ src/app/api/order/route.ts
@@ -2,8 +2,7 @@ import { responseHelper } from '@/lib/helpers';
 import { isAuthenticatedAndUserData } from '@/lib/auth';
 import prisma from '@/lib/prisma/client';
 import config from '@/config';
-import { directus } from '@/lib/utils';
-import { createItems } from '@directus/sdk';
+import { Order } from '@/types/server/types';
 
 // TODO Get detail of order by order id
 export async function GET(req: Request) {
@@ -30,56 +29,72 @@ export async function GET(req: Request) {
 
 export async function POST(req: Request) {
     try {
-        
         const auth = await isAuthenticatedAndUserData();
         const userId = auth.user?.id;
+        const { addressId, paymentOption, fullAddress, cartTotal } = await req.json();
 
         if (!userId) {
             return responseHelper({ message: 'User not authenticated', statusCode: 401, data: {} }, 401);
         }
 
-
         const detailedOrder: any[] = await prisma.$queryRaw`
-        select c.user_id, c.product_id, c.cart_quantity, c.quantity_id as "cart_to_quantiy_ref",
-        q.price as "price_per_unit", q.quantity as "product_quantity", q.price * c.cart_quantity as "amount",
-        p.name as "product_name", p.description, 
-        c.created_at, c.updated_at
+        select c.user_id, c.product_id, c.cart_quantity, q.quantity,
+        q.price, p.name , p.description
         from cart as c
         join quantity as q on c.quantity_id = q.id
         join products as p on c.product_id = p.id
-        where c.user_id = ${userId}
-        `
+        where c.user_id = ${userId}`
 
-        const totalAmount = detailedOrder.reduce((acc, curr) => acc + Number(curr.amount), 0);
+        const totalAmount = detailedOrder.reduce((acc, curr) => acc + curr.cart_quantity * Number(curr.price), 0);
 
-        const newOrder = {
-            user_id: userId,
-            total_amount: totalAmount,
-            status: config.order_status.ORDER_PLACED,
-            order_date: new Date().toISOString(),
-            created_at: new Date().toISOString(),
-            updated_at: new Date().toISOString(),
-        };
-        const newOrderCreated = await prisma.orders.create({ data: newOrder });
+        if (totalAmount !== cartTotal) {
+            return responseHelper({ message: 'Cart total does not match', statusCode: 400, data: {} }, 400);
+        }
 
-        const orderItems = detailedOrder.map((item) => {
-            return {
-                order_id: newOrderCreated.id,
+        const newOrder = await prisma.$transaction(async (prisma) => {
+            const order: Order[] = await prisma.$queryRaw`
+                INSERT INTO orders (user_id, total_amount, status, order_date, created_at, updated_at, address_id, full_address)
+                VALUES (${userId}, ${totalAmount}, ${config.order_status.ORDER_PLACED}, NOW(), NOW(), NOW(), ${Number(addressId)}, ${fullAddress})
+                RETURNING *
+            `;
+            console.log(order, 'order');
+            // return;
+            const orderId: number = order[0].id;
+            const orderItems = detailedOrder.map((item) => ({
+                order_id: orderId,
                 product_id: item.product_id,
-                product_name: item.product_name,
-                product_quantity: item.product_quantity,
+                product_name: item.name,
+                product_quantity: item.quantity,
                 quantity: item.cart_quantity,
-                price_per_unit: item.price_per_unit,
-                created_at: new Date().toISOString(),
-                updated_at: new Date().toISOString(),
-            }
-        })
+                price_per_unit: item.price,
+                created_at: new Date(),
+                updated_at: new Date(),
+            }));
 
-        const orderItemsCreated = await directus.request(createItems('order_items', orderItems));
+            const orderItemsCreated = await prisma.order_items.createMany({
+                data: orderItems
+            });
 
-        await prisma.cart.deleteMany({ where: { user_id: userId } });
+            await prisma.payments.create({
+                data: {
+                    user_id: userId,
+                    order_id: orderId,
+                    payment_date: new Date(),
+                    amount: totalAmount,
+                    payment_method: paymentOption,
+                    created_at: new Date(),
+                    updated_at: new Date(),
+                }
+            });
 
-        return responseHelper({ message: 'Order placed successfully', statusCode: 200, data: orderItemsCreated }, 200);
+            await prisma.cart.deleteMany({ where: { user_id: userId } });
+
+            return order;
+        });
+
+        
+
+        return responseHelper({ message: 'Order placed successfully', statusCode: 200, data: newOrder }, 200);
     } catch (err) {
         console.error('Internal server error:', err);
         return responseHelper({ message: 'Internal server error', statusCode: 500, data: {} }, 500);
